image displays + bug fix
added image displays with full functionality aside from click to edit (coming soon).
i actually added the landscape and fit options in this update, and backported them to android.
fit can be set to fit (default), zoom, or stretch, and controls how the image is resized to fit in the bounding box.

fixed a bug where landscape buttons were using the wrong hitbox for click detection.