added radio buttons
added radio buttons and added the usual fontSize/landscape/readonly settings.
backported these updates to Android, as usual.

this completes the pre-existing custom controls, so controls are caught up to Android now.
the only thing that's not fully working for controls is that image boxes don't support the camera yet (soon).

after this, we just need to work on some of the more oddball sensors and we'll be done.