added camera support for image displays
added in the camerea support stuff for editing image displays.
thankfully, iOS makes this much less of a headache than it was in Android.

fixed a bug where we were forgetting to send update events for text fields.
added the update events for image displays (weren't present before because you couldn't edit them).