background network close bug fix
apparently iOS closes network connections immediately upon going into the background.
this includes both going to home screen, as well as locking the device.
thus, after these events, networking was no longer working.
i fixed it so that it will reconnect after a failure.
unfortunately, this also means (unless we can get around this limitation) we can't do a runInBackground feature like the Android version has.

implemented everything needed for a runInBackground feature.
mostly useless now due to above, but it's there in case we can fix the problem later.
the scenePhase part at least is useful either way, as it was needed to do the reconnect logic.