unit fixes and hex str patch
fixed an issue where we were sending the wrong units over to the server for rotation based controls.
we now uniformly send degrees and the server doesn't have to do anything at all.

fixed a cosmetic issue with watchers in NetsBlox where hex strings that happen to take the form of exponential notation display as a number.
this isn't a problem for us, but it's easy to avoid at the source.