visual fixes
things were hard to see in dark theme, so i just locked the app to light theme.
since custom control colors come from NetsBlox, we can't always avoid black on black.
so it's better to just have the app always look the same.

apparently enabling the proximity sensor causes the display to turn off when it's triggered.
there's no way to prevent this, so i've elected to just not support proximity.
it's super close range anyway, so not very interesting as a sensor.